INSP: merge RsMissingFeaturesInspection and CargoTomlMissingFeaturesInspection
As a result, Rust inspection settings contain only one `Missing features` inspection instead of two identical items
