--- toml/src/main/resources/inspectionDescriptions/CargoTomlMissingFeatures.html
@@ -1,5 +1,0 @@
-<html>
-<body>
-Detects Cargo packages with missing cargo features in some of their dependency packages
-</body>
-</html>
