+++ toml/src/test/kotlin/org/rust/toml/inspections/MissingFeaturesInspectionTest.kt
@@ -3,8 +3,9 @@
  * found in the LICENSE file.
  */
 
-package org.rust.ide.inspections
+package org.rust.toml.inspections
 
+import com.intellij.openapi.application.runWriteAction
 import com.intellij.testFramework.InspectionTestUtil
 import org.rust.FileTree
 import org.rust.cargo.RsWithToolchainTestBase
@@ -15,7 +16,8 @@ import org.rust.fileTree
 import org.rust.singleProject
 import org.rust.workspaceOrFail
 
-class RsMissingFeaturesInspectionTest : RsWithToolchainTestBase() {
+class MissingFeaturesInspectionTest : RsWithToolchainTestBase() {
+
     fun `test missing dependency feature`() = doTest(
         fileTree {
             toml("Cargo.toml", """
@@ -35,9 +37,7 @@ class RsMissingFeaturesInspectionTest : RsWithToolchainTestBase() {
                 """)
                 dir("src") {
                     file("main.rs", """
-                        <warning descr="Missing features: bar/feature_bar">
                         fn main() {}
-                        </warning>
                     """)
                 }
             }
@@ -82,9 +82,7 @@ class RsMissingFeaturesInspectionTest : RsWithToolchainTestBase() {
             """)
             dir("src") {
                 file("main.rs", """
-                    <warning descr="Missing features: hello/feature_hello">
                     fn main() {}
-                    </warning>
                 """)
                 file("lib.rs", "")
             }
@@ -92,6 +90,50 @@ class RsMissingFeaturesInspectionTest : RsWithToolchainTestBase() {
         pkgWithFeature = "hello",
         featureName = "feature_hello",
         fileToCheck = "src/main.rs"
+    )
+
+    fun `test missing dependency feature in manifest`() = doTest(
+        fileTree {
+            toml("Cargo.toml", """
+                [workspace]
+                members = ["foo", "bar"]
+            """)
+
+            dir("foo") {
+                toml("Cargo.toml", """
+                    [package]
+                    name = "foo"
+                    version = "0.1.0"
+                    authors = []
+
+                    [dependencies]
+                    bar = { path = "../bar", features = ["feature_bar"] }
+                """)
+                dir("src") {
+                    file("main.rs", """
+                        fn main() {}
+                    """)
+                }
+            }
+
+            dir("bar") {
+                toml("Cargo.toml", """
+                    [package]
+                    name = "bar"
+                    version = "0.1.0"
+                    authors = []
+
+                    [features]
+                    feature_bar = [] # disabled
+                """)
+                dir("src") {
+                    file("lib.rs", "")
+                }
+            }
+        },
+        pkgWithFeature = "bar",
+        featureName = "feature_bar",
+        fileToCheck = "foo/Cargo.toml"
     )
 
     private fun doTest(tree: FileTree, pkgWithFeature: String, featureName: String, fileToCheck: String) {
@@ -103,10 +145,15 @@ class RsMissingFeaturesInspectionTest : RsWithToolchainTestBase() {
 
         project.cargoProjects.modifyFeatures(cargoProject, setOf(PackageFeature(pkg, featureName)), FeatureState.Disabled)
 
-        val enabledInspections = InspectionTestUtil.instantiateTool(RsMissingFeaturesInspection::class.java)
-        myFixture.enableInspections(enabledInspections)
+        val enabledInspection = InspectionTestUtil.instantiateTool(MissingFeaturesInspection::class.java)
+        myFixture.enableInspections(enabledInspection)
 
         myFixture.openFileInEditor(cargoProjectDirectory.findFileByRelativePath(fileToCheck)!!)
+        runWriteAction {
+            with(myFixture.editor.document) {
+                setText("<warning descr=\"Missing features: $pkgWithFeature/$featureName\">$text</warning>")
+            }
+        }
         myFixture.checkHighlighting(
             /* checkWarnings = */ true,
             /* checkInfos = */ false,
